network/requirements.txt
albumentations>=1.3.1
matplotlib>=3.7.0
pandas>=2.0.0
pillow>=9.0.0
pytorch-lightning>=2.0.0
scikit-learn>=1.2.0
timm>=0.9.2
torch>=2.0.0
torchvision>=0.15.0
